Move pages to pages lib
