+++ apps/digital-www-pwa/src/app/events/[day]/page.tsx
@@ -1,27 +1,7 @@
 'use client';
 
-import { useFeedContext } from '@digital-www-pwa/feed-context';
+import { EventsPage } from '@digital-www-pwa/pages';
 
 export default function Page() {
-  const feed = useFeedContext();
-  return (
-    <>
-      <header className="bg-white shadow">
-        <div className="mx-auto max-w-7xl px-4 py-6 sm:px-6 lg:px-8">
-          <h1 className="text-3xl font-bold tracking-tight text-gray-900">
-            Events
-          </h1>
-        </div>
-      </header>
-      <main>
-        <div className="mx-auto max-w-7xl px-4 py-6 sm:px-6 lg:px-8">
-          <ul>
-            {feed.coalesce.map((event) => {
-              return <li key={event.event_id}>{event.title}</li>;
-            })}
-          </ul>
-        </div>
-      </main>
-    </>
-  );
+  return <EventsPage />;
 }
