+++ apps/digital-www-pwa/src/app/favorites/[day]/page.tsx
@@ -1,18 +1,6 @@
+'use client';
+import { FavoritesPage } from '@digital-www-pwa/pages';
+
 export default function Page() {
-  return (
-    <>
-      <header className="bg-white shadow">
-        <div className="mx-auto max-w-7xl px-4 py-6 sm:px-6 lg:px-8">
-          <h1 className="text-3xl font-bold tracking-tight text-gray-900">
-            Favorites
-          </h1>
-        </div>
-      </header>
-      <main>
-        <div className="mx-auto max-w-7xl px-4 py-6 sm:px-6 lg:px-8">
-          Favorites Page
-        </div>
-      </main>
-    </>
-  );
+  return <FavoritesPage />;
 }
