+++ libs/pages/src/index.ts
@@ -1,3 +1,10 @@
 // Use this file to export React client components (e.g. those with 'use client' directive) or other non-server utilities
 
-export * from './lib/pages';
+export * from './lib/home';
+export * from './lib/art';
+export * from './lib/camps';
+export * from './lib/events';
+export * from './lib/favorites';
+export * from './lib/radio';
+export * from './lib/vehicles';
+export * from './lib/week-view';
