--- libs/pages/src/lib/pages.module.css
@@ -1,7 +1,0 @@
-/*
- * Replace this with your own classes
- *
- * e.g.
- * .container {
- * }
-*/
