--- libs/pages/src/lib/pages.tsx
@@ -1,11 +1,0 @@
-import styles from './pages.module.css';
-
-export function Pages() {
-  return (
-    <div className={styles['container']}>
-      <h1>Welcome to Pages!</h1>
-    </div>
-  );
-}
-
-export default Pages;
