+++ libs/pages/src/lib/radio/index.tsx
@@ -1,0 +1,20 @@
+'use client';
+
+export function RadioPage() {
+  return (
+    <>
+      <header className="bg-white shadow">
+        <div className="mx-auto max-w-7xl px-4 py-6 sm:px-6 lg:px-8">
+          <h1 className="text-3xl font-bold tracking-tight text-gray-900">
+            Radio
+          </h1>
+        </div>
+      </header>
+      <main>
+        <div className="mx-auto max-w-7xl px-4 py-6 sm:px-6 lg:px-8">
+          Radio Page
+        </div>
+      </main>
+    </>
+  );
+}
